+++ action_server/src/sema4ai/action_server/_download_rcc.py
@@ -8,7 +8,7 @@ from typing import Optional
 log = logging.getLogger(__name__)
 
 # Note: also referenced in action_server/build.py
-RCC_VERSION = "17.23.2"
+RCC_VERSION = "17.28.4"
 RCC_URLS = {
     "Windows": f"https://sema4.ai/cdn/downloads/rcc/releases/v{RCC_VERSION}/windows64/rcc.exe",
     "Darwin": f"https://sema4.ai/cdn/downloads/rcc/releases/v{RCC_VERSION}/macos64/rcc",